// FCAPROGRAMACION/FrontEnd/FCAPROGFE001/src/app/models/Produccion/Paros.ts
export interface Paros {
  claveParo: string;
  tipoParo: string;
  tipoParoValida: string;
  tipoMaquina: string;
  aplicaPareto: boolean;
  descuentoIndicador: boolean;
  observaciones: string;
  autoriza: boolean;
  esProgramado: boolean;
  estatus: boolean;
}
export class Paros {
  claveParo: string;
  tipoParo: string;
  tipoParoValida: string;
  tipoMaquina: string;
  aplicaPareto: boolean;
  descuentoIndicador: boolean;
  observaciones: string;
  autoriza: boolean;
  esProgramado: boolean;
  estatus: boolean;
}